Restore HDD folder creation
Create HDD by default, then copy files to the disk image.
May be removed later, as the whole point of creating a disk image is to have multiple partitions.
Just using this method for now as it is faster to test with, and we are not using the extra partition "OS" on the disk image.